Moved the server into its proper class and added Groovy and Java based testing.
