+++ TODO.md
@@ -1,4 +1,6 @@
 
+- server should be restartable (to apply additional configuration)
+- need parallel testing of the Java builder API
 
 - setup Travis for CI builds
 - setup Coveralls for coverage
