+++ build.gradle
@@ -25,6 +25,7 @@ dependencies {
     compile 'org.codehaus.groovy:groovy-all:2.4.6'
     compile 'io.undertow:undertow-core:1.4.6.Final'
 
+    testCompile 'com.squareup.okhttp3:okhttp:3.5.0'
     testCompile 'org.spockframework:spock-core:1.0-groovy-2.4'
 
     asciidoclet 'org.asciidoctor:asciidoclet:1.+'
