+++ src/main/java/com/stehno/ersatz/ErsatzServer.java
@@ -1,4 +1,23 @@
+/**
+ * Copyright (C) 2016 Christopher J. Stehno
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.stehno.ersatz;
+
+import io.undertow.Undertow;
+import io.undertow.server.HttpHandler;
+import io.undertow.server.HttpServerExchange;
 
 import java.util.function.Consumer;
 
@@ -13,7 +32,25 @@ public class ErsatzServer {
         expects.accept(expectations);
     }
 
-    public ServerExpectations getExpectations() {
-        return expectations;
+    // FIXME: should be restartable
+    void start(){
+        Undertow server = Undertow.builder()
+            .addHttpListener(8080, "localhost")
+            .setHandler(new HttpHandler() {
+                @Override
+                public void handleRequest(final HttpServerExchange exchange) throws Exception {
+                    expectations.requests().forEach(getreq -> {
+                        if (matches(exchange, getreq)) {
+                            exchange.getResponseSender().send(getreq.getResponse().getBody().toString());
+                        }
+                    });
+                }
+            }).build();
+
+        server.start();
+    }
+
+    private static boolean matches(final HttpServerExchange exchange, final GetRequest request) {
+        return exchange.getRequestPath().equals(request.getPath());
     }
 }
