+++ src/test/groovy/com/stehno/ersatz/ErsatzServerTest.java
@@ -1,0 +1,44 @@
+/**
+ * Copyright (C) 2016 Christopher J. Stehno
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.stehno.ersatz;
+
+import okhttp3.OkHttpClient;
+import okhttp3.Request;
+import org.junit.Test;
+
+import java.io.IOException;
+
+import static org.junit.Assert.assertEquals;
+
+public class ErsatzServerTest {
+
+    @Test
+    public void prototype() throws IOException {
+        final ErsatzServer ersatzServer = new ErsatzServer();
+
+        ersatzServer.requesting(expectations -> {
+            expectations.get("/foo").responds().body("This is Ersatz!!");
+            expectations.get("/bar").responds().body("This is Bar!!");
+        });
+
+        ersatzServer.start();
+
+        OkHttpClient client = new OkHttpClient();
+        Request request = new Request.Builder().url("http://localhost:8080/foo").build();
+
+        assertEquals("This is Ersatz!!", client.newCall(request).execute().body().string());
+    }
+}
